// src/types/feedback.ts
export type FeedbackType = "誤字脱字" | "正誤情報" | "読みにくい" | "その他";
export type StatusType =
  | "未対応"
  | "出版社へ送信済み"
  | "出版社確認中"
  | "出版社回答済み"
  | "修正予定"
  | "修正済み"
  | "対応なし"
  | "却下";
export type LocationType = "ページ" | "Kindle" | "その他";

export type Feedback = {
  id: string;
  bookId: string;
  userId: string;
  userName: string;
  userIdShort: string;
  title: string;
  bookTitle: string;
  bookAuthor: string;
  publisher: string;
  isbn: string;
  coverImage: string;
  edition?: number;
  printing?: number;
  type: FeedbackType;
  locationType: LocationType;
  page?: number;
  line?: number;
  hasMultiplePages?: boolean;
  locationNote?: string;
  kindleLocation?: string;
  wrong?: string;
  correct?: string;
  content?: string;
  note?: string;
  publisherComment?: string;
  status: StatusType;
  fixedEdition?: number;
  fixedPrinting?: number;
  createdAt: string;
};